server/pyproject.toml
[tool.poetry]
name = "raspberry-pi-shop"
version = "0.1.0"
description = ""
authors = ["Atanas Pozharliev <ABPozharliev19@codingburgas.bg>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.2"
psycopg2 = "^2.9.4"
djangorestframework-simplejwt = "^5.2.2"
azure-storage-blob = "^12.14.1"
azure-identity = "^1.12.0"
pillow = "^9.4.0"
    
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"